fix: manual backport to SQL view in other path [DHIS2-9213]
